desktop/requirements.txt
numpy==1.22.4
Pillow==9.1.1
plyer==2.0.0
requests==2.27.1
ttkwidgets==0.12.1
